+++ src/combinators/prepend.test.js
@@ -1,18 +1,18 @@
 import Signal from '../Signal'
 import concat from './concat'
-import startWith from './startWith'
+import prepend from './prepend'
 
 jest.mock('../Signal')
 jest.mock('./concat')
 
-describe('startWith', () => {
+describe('prepend', () => {
   it('calls concat', () => {
     const s = 'foo'
     const t = 'bar'
 
     Signal.of.mockReturnValue(t)
 
-    startWith(0, s)
+    prepend(0, s)
 
     expect(Signal.of).toHaveBeenCalledWith(0)
     expect(concat).toHaveBeenCalledWith(t, s)